+++ nautilus_trader/model/instruments/futures_contract.pyx
@@ -68,6 +68,8 @@ cdef class FuturesContract(Instrument):
         The UNIX timestamp (nanoseconds) when the data event occurred.
     ts_init : uint64_t
         The UNIX timestamp (nanoseconds) when the data object was initialized.
+    exchange : str, optional
+        The instruments ISO 10383 Market Identifier Code (MIC).
     info : dict[str, object], optional
         The additional instrument information.
 
@@ -81,6 +83,8 @@ cdef class FuturesContract(Instrument):
         If `tick_size` is not positive (> 0).
     ValueError
         If `lot_size` is not positive (> 0).
+    ValueError
+        If `exchange` is not ``None`` and not a valid string.
     """
 
     def __init__(
@@ -98,8 +102,11 @@ cdef class FuturesContract(Instrument):
         uint64_t expiration_ns,
         uint64_t ts_event,
         uint64_t ts_init,
+        str exchange = None,
         dict info = None,
     ):
+        if exchange is not None:
+            Condition.valid_string(exchange, "exchange")
         super().__init__(
             instrument_id=instrument_id,
             raw_symbol=raw_symbol,
@@ -127,6 +134,7 @@ cdef class FuturesContract(Instrument):
             ts_init=ts_init,
             info=info,
         )
+        self.exchange = exchange
         self.underlying = underlying
         self.activation_ns = activation_ns
         self.expiration_ns = expiration_ns
@@ -138,6 +146,7 @@ cdef class FuturesContract(Instrument):
             f"raw_symbol={self.raw_symbol}, "
             f"asset_class={asset_class_to_str(self.asset_class)}, "
             f"instrument_class={instrument_class_to_str(self.instrument_class)}, "
+            f"exchange={self.exchange}, "
             f"quote_currency={self.quote_currency}, "
             f"underlying={self.underlying}, "
             f"activation={format_iso8601(self.activation_utc)}, "
@@ -196,6 +205,8 @@ cdef class FuturesContract(Instrument):
             expiration_ns=values["expiration_ns"],
             ts_event=values["ts_event"],
             ts_init=values["ts_init"],
+            exchange=values["exchange"],
+            info=values.get("info"),
         )
 
     @staticmethod
@@ -220,6 +231,8 @@ cdef class FuturesContract(Instrument):
             "margin_maint": str(obj.margin_maint),
             "ts_event": obj.ts_event,
             "ts_init": obj.ts_init,
+            "exchange": obj.exchange,
+            "info": obj.info,
         }
 
     @staticmethod
@@ -238,6 +251,7 @@ cdef class FuturesContract(Instrument):
             expiration_ns=pyo3_instrument.expiration_ns,
             ts_event=pyo3_instrument.ts_event,
             ts_init=pyo3_instrument.ts_init,
+            exchange=pyo3_instrument.exchange,
         )
 
     @staticmethod